8.1.2 login form
This reuses the markup pattern for the sign-up form. The tutorial has
made a distinction between signing up creating poets (users) and loging
in creating sessions. This gives a clue to how I can think about the
markup. The two forms are for different types of data, and so should
have different markup. Different combinations of body, form, fieldset and
section ought to work.