8.3 ex 1, 2 Logging out
Works ok. I found a bug (that I'd created): the log out path was wrong
(session#new rather than session#destroy) in the router. I thought about
writing a test to catch this, but remembered that I found the problem
initially by running the test created in the tutorial - i.e. it already
was tested.

The final exercise instructs to inspect the browser cookies to ensure
the session has been nil'd properly. This seems hard to confirm: the
session cookie is encrypted. The byte count changes, so that'll do for
me.