10.3.2 ex 1 Sample Poets
- Going by the tutorial, I'd added 'settings' to the "poem navigation'
  (the top nav bar). This tricked me into think it would edit the
settings for the current *viewed" poet, but it doesn't. Instead, it
links to the edit page for the current logged in poet. That makes sense,
and the tutorial had described that already... however, it shows up a
usability quirk, or a quirk of my psychology: I was expecting the
controls to act on what I could see, not what state the application
thought it was in. In part, this expectation will likely stem from my
drive to have the edit mode look like the read mode as much as possible
(to make the site WYSYWYG), but I wonder if that itself is a reflection
of my "edit in place" vibe, or even if it's a reflection of how most
people view these sorts of things. Word processors and spreadsheets have
us conditioned to think about edit in place too.. sorta.
- that it feels risky adding loads of fake users to the seeds, in
  production. I need to see this in practice I think...
- faker gem looks fun