13.3.2 Creating poems - going off-piste
The tutorial is pushing micropost creation behind login, and show on
profile pages. My poems should be added wether logged in or not. They
will be shown on poet's pages, but there's a bit of uncertainty right
now about the structure for that (is it the logged-out view of a poet,
or an index view of poems? or both?)

Currently, I've broken some tests. In a strange way.