13.3.2 Creating poems - fixed a bug in the tests
Tests were failing intermittently. So frustrating! The fail was that
certain poems weren't found on the poet profile page. At first I thought
this was an ordering problem (and discovered that default_scope is
largely frowned upon). Checking the test output more closely I realised
it was an escaping issue. Model strings get escaped when written itno
the HTML so & and ' become &amp; and &#39; respectively. The **tests**
don't escape the model data when passing it to assert_match, so some of
the poems wouldn't match despite appearing on the page in human terms.
The fix was to use ERB::Util.h to escape the model property before
handing it to assert_match.